Move lalinference modules from lalapps to lalinference/python
