+++ lalapps/src/inspiral/posterior/lalapps_nest2pos.py
@@ -5,9 +5,10 @@ import h5py
 import os
 import sys
 
+import lalinference
 from lalinference import LALInferenceHDF5PosteriorSamplesGroupName as posterior_grp_name
 from lalinference import LALInferenceHDF5NestedSamplesGroupName as nested_grp_name
-from lalapps.nest2pos import draw_posterior_many, draw_N_posterior_many, compute_weights
+from lalinference.nest2pos import draw_posterior_many, draw_N_posterior_many, compute_weights
 
 usage = '''%prog [-N Nlive] [-p posterior.hdf5] [-H header.txt] [--npos Npos] datafile1.hdf5 [datafile2.hdf5 ...]
